+++ Sources/NSManagedObject+Utils.swift
@@ -74,4 +74,11 @@ extension NSManagedObject {
   public final func delete() {
     managedObjectContext?.delete(self)
   }
+
+  /// **CoreDataPlus**
+  /// Converts the object ID to a permanent ID (if the object doesn't already have a permanent ID)
+  public func obtainPermanentID() throws -> NSManagedObjectID  {
+    try managedObjectContext?.obtainPermanentIDs(for: [self])
+    return self.objectID
+  }
 }